chore: move tsToJS to utils.ts
